requirements.txt
cython
h5py
tqdm
msgpack
pillow
imageio
numpy
scipy
scikit-image
gdown
matplotlib
natsort
Pillow
PyYAML==6.0
scikit-image==0.18.3
scipy==1.7.1
setuptools
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
timm
torch==1.9.0+cu102
torchaudio==0.9.0
torchnet==0.0.4
torchvision==0.10.0+cu102